nassh: fs: add logic to sync indexeddb-fs & DOM FS when running NaCl
The NaCl plugin is able to read the DOM FS directly (which is what
made it an attractive option in the first place).  Now that we're
moving backing storage to indexeddb-fs, we need to manually sync
the state to the DOM FS before launching the NaCl plugin, and then
pull back in updates the NaCl plugin might have made when it exits.
Currently the only file that OpenSSH is known to write to is the
known_hosts file, so syncing it by hand shouldn't be terrible.

Bug: https://crbug.com/1112490
Bug: https://crbug.com/1298699
Change-Id: I3fd771718ed9863de97d7daddab46c9061d4356b
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3780840
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jason Lin <lxj@google.com>