libdot: fix preference manager defaults for objects
When a preference is an object (array or object), we must always return
a copy of the object to ensure that any changes do not modify the value
managed by PreferenceManager.

Fixes diff() so that when setting a value, we can correctly detect when
the new value matches the default and then delete all storage.

Bug: 1330425
Change-Id: I1b3972ff66520e657f389681f66ebf4b18fae227
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3676614
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>