terminal: leave out ssh settings_profile param for default
Only inclue terminal_ssh.html?settings_profile=<not-default> when
profile is not 'default'.  This is needed as a short term fix to ensure
that chrome.runtime.connect() will be provided to SSH, but the change is
OK to leave longer term.  See crrev.com/c/3690189 for fix to allow
chrome.runtime.connect() to be included even when URL params exist.

Bug: 1330425
Change-Id: I4712bc96e08e9ede2c02727ad63ab18fbff574e7
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3686955
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>