libdot: pylint: leverage vpython when available
This allows us to pin a specific pylint version if the developer has
vpython available.  If they don't, we still fallback to the system
copy so we work with external developers.

Since kokoro will use the newer pinned version of pylint now, fold a
few warning fixes in here that wouldn't work with the system version
of pylint that's currently in the (Debian Buster) container.

Change-Id: Ia2c70de5efa6bd161c80cdb6036df1b3872a3ed2
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3783661
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>