nassh: mkdist: update list of files
Drop the wassh/test* files now that the wassh port is further along.
They're still available to us when developing locally.

Add the wasm port to the crosh build so it's included for Terminal
in CrOS releases.

Also tweak the way we pull in plugin files to omit the plugin/.hash
file since it isn't needed by anything.

Change-Id: Id11f33abf64e3cccf1d22a5f1dcc07920777bda2
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3795333
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>