nassh: fix corp relay v4 reconnect
Remove old callbacks and delay creating a new WebSocket for 2s.
During CrOS lid-close-suspend, the network stack goes a little strange
(I guess that is to be expected).  Existing sockets are closed and
onclose events are fired, but if we immediately create a new WebSocket,
it will sometimes get started before suspend completes.  It is much
better to delay and ensure we do not start a new WebSocket until
lid-open-resume.

Modified reconnect logic to set a limit of 3 failed attempts at
reconnect. Previously we could get in a bad state where our session is
removed from GRUE, and we do an endless loop of reconnects.

Bug: 1314801

Change-Id: I0c43509af85ce4e14570c79095f631575b7a3c95
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3812955
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>