libdot: dist: centralize logic
In preparation for turning libdot into an ES6 module, move all of the
dist logic to libdot itself.  This simplifies hterm & nassh a bit at
the cost of hterm's dist including a few extra modules, but that only
adds ~8k which is nothing when all of libdot is ~174k.  Once we switch
to an ES6 module, this would already happen, but it would get fixed if
we enable tree shaking (which would clean even more).

This also fixes a minor bug where we were including lib_codec.js twice
in nassh (once via hterm, then via nassh).

Bug: https://crbug.com/1350884
Change-Id: I7c9d89d0eb4b76ea3130d05cba00a71e3fad8145
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3932216
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>