nassh: relays: harmonize onWriteSuccess usage
We have 3 different styles of testing the onWriteSuccess callback in
our 5 relay implementations.  Harmonize on just 1 style since there's
no point in deviating.

In practice this affects only 1 scenario: SSH-FE with WASM where the
WASM code calls asyncWrite w/out a callback.  In that case, we set it
to undefined, and the later relay check to make the call iff it isn't
null fails because undefined isn't a function.

Bug: b/246487510
Change-Id: Ibbabfe62c512886c2060f18391a21a279f2385a3
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3894089
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>