Use correct write count for ssh-agent relay
Plugin OnWriteAcknowledged() expects the total running write count,
rather than the size of the latest packet to be written.

We set the default writeWindow to 8K in nassh_command_instance.
The plugin sends 2 packets quickly to the relay with key data, and
something else about 2K.  When key data gets to about 8K size, the first
packets sends OK and gets acked with 8K count, but after sending the 2nd
packet, and getting count of 2K, plugin becomes stuck.

Bug: b/140365615
Change-Id: If0a5977e67d3991f85bfd442fb9526ab184a2167
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3905589
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>