wassh: readpassphrase: handle ONLCR when displaying prompt
When printing strings with readpassphrase, make sure we handle ONLCR
(insert \r when we see \n) so output strings are properly aligned.

This should probably be in the tty layer instead, but that doesn't
seem to be necessary in general with OpenSSH, and this is simple
enough to make things work for users (practically speaking).

Bug: https://crbug.com/1312160
Change-Id: Idd9de65a74798d9cc19605818439120edf27f845
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3916053
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>