terminal: stop normalizing font family and size
Stop auto-normalizing (correcting) font-family and font-size. This code
was originally added when we expected to be importing prefs from
nassh, but we are no longer doing that for new users.

If a user overrides font-family or font-size by some unsupported means,
we can display the value they have chosen, and allow them to use the UI
to select any new supported value, but we will not normalize back to the
default value.

Bug: 1050883

Change-Id: I127f195fb52d6acb0c58d4b6f30ceee8c1994b0e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3924006
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>