wassh: fs: fix fd exhaustion bug
The FdMap class handling of this.next_ was lazy.  We always incremented
it when searching for the next free fd and never decremented it.  This
is simpler (as we don't have to search for holes in the map), and it can
make debugging easier (when looking at logs, you know a fd will never be
reused).  However, OpenSSH uses select() on its fds, and that has a max
value of 1024.  So as soon as we have a long enough running process that
keeps opening & closing new fd's (e.g. lots of ssh-agent traffic), this
eventually crashes OpenSSH when we try to exceed the max.

Fix it by setting a max fd in FdMap and having next() loop around to the
start of the address (fd) space when it hits the end.

Bug: https://crbug.com/1369026
Change-Id: I4818d2543c287883d60a40340b083356d8093fb7
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3924011
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Emil Mikulic <easy@google.com>