wassh: sockets: handle -4/-6 preferences with Chrome TCP APIs
This updated API exists in R103, so should be safe to use now.  The
WASM port requires newer versions anyways, and we'll probably bump
it before we release.

We have to roll the Chrome externs to the very latest to pick up this
new API as it was only updated very recently.

Bug: https://crbug.com/1311912
Change-Id: I1ab0a0a02925df989d032c065b8f029dfa7077c3
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3928575
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Emil Mikulic <easy@google.com>