libdot: convert window to globalThis
The window object isn't available in service workers.  globalThis
should be available everywhere, so switch all code over to that.

There are some places where this is not required, but seems better
to be consistent since globalThis always works rather than try and
figure out which contexts the code runs in (or might in the future).

Bug: https://crbug.com/1350884
Change-Id: I2d4e1f719b1410d13c102e722622c7a8159ac3a8
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3935495
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>