nassh: rollup: split deps apart
We don't want all of these deps in the same modules.  This hasn't been
an issue so far, but it's turning out that some of these modules don't
work in service workers.  For example, the UI modules (e.g. Lit) want
HTML web APIs.

Let's split these up into unrelated projects so we only import them
when they're used.

The final size of the individual outputs is the same as the single one
since many of these don't overlap.  Along those lines, group the few
together that are related so we don't duplicate the imports.

Bug: https://crbug.com/1350884
Change-Id: Idf32cda07f0e37e6f1578aa95f8cc0fec67ff5af
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3935676
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>