nassh: rename nassh_deps.concat.js to deps_local.concat.js
Use a new deps_xxx namespace to hold the increasing number of 3rd party
dependencies we're pulling in, and to help differentiate from the 1st
party sources we maintain.  We'll use this more in a follow up commit.

We use $PROJECT_ as the prefix across all our projects in general when
we intend to release them standalone (e.g. libdot & hterm).  That isn't
exactly the same with nassh where we release it by itself, even though
we are flattening the namespace with terminal/ currently.

Bug: https://crbug.com/1350884
Change-Id: I04016699eb7838de115fc5bbfa5e3ba1baa09e6f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3937285
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>