nassh: prefs: switch local prefs to chrome.storage.local
We've been using localStorage to store local preferences, and to share
them between the foreground (connections) and background pages.  With
MV3, we have to switch the background page to a service worker, and
those don't have access to localStorage APIs.  Let's switch over to
Chrome's storage.local APIs as all contexts have access to it.

Bug: https://crbug.com/1350884
Change-Id: I0da29d6ce51e5a36d22b1f46453dfb46a56a3bd8
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3963876
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>