nassh: connect: switch extra fields to a table for alignment
Rather than use a custom x-box to align each sub-field with the overall
dialog and then manually compute the label widths to align them, switch
to a table to let the browser align everything for us.  We put the table
in an x-box to keep the overall table aligned with the other fields.

Change-Id: I38b0ab267b6fddb56f73ff048818baa79f5be3c6
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4717025
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>