pull default location from globalThis (i.e. window)
Since the location object is in global scope, use that instead of
the one attached to the document in case we don't have access to
the document object (such as a worker thread).

Further, this is considered best practice as window.location has
always been the current page location, while document.location was
a string before changing to a read-only copy of window.location.

https://html.spec.whatwg.org/multipage/nav-history-apis.html#dom-location

Change-Id: I3717eb09042acf734f2cc68917fc4395c9f239c0
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4691789
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>