nassh: dist: stop mangling the dev extension name
We rely on the version_name field now to detect whether the extension
is a "developer" build, so stop mangling the version string.  CWS can't
handle partially localized strings.

While this means the extension name in the UI is just "Secure Shell",
we use a different icon & version field to indicate it's not the stable
version.  Not great, but hopefully it's still visually distinct enough
for people to flip between the two.

Bug: b/153099839
Change-Id: I2b42433adb98251a95f8128347f9c7c5321e8c7d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4690633
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>