node_modules: refresh a ton of packages
Update all the packages to lint & build.  Packages used at runtime are
left alone for now.  Also add @rollup/plugin-yaml to the mix as we'll
use it in nassh for resource loading, and rollup-plugin-concat to help
replace our custom concat script, and rollup-plugin-string to help with
importing SVG files as plain strings.

Change-Id: Iff70bdfa0a2e059d711f9528a9f41667c1b7cbd3
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4800519
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>