ssh_client: openssh: increase fd_set size to 4096
When we added FD_SETSIZE checking and fixed the allocation logic for
wasm, we ended up scaling back the NaCl side to 64 since that's what
the NaCl C library uses.  This ends up being too small for some users,
so increase this to 4k under NaCl.

Normally this wouldn't work that well, but since OpenSSH is the only
thing allocating this structure, and it only passes it to our custom
plugin APIs, there shouldn't be any other place where the structure
layout would be out of sync.

Bug: b/260531419
Test:
  run Secure Shell with -R localhost:9000:localhost:9001
  run iperf server: iperf -s -p 9001
  run iperf client: iperf -p 9000 -P 10 --connect-only=0 -c localhost
Change-Id: I8ca016822886a6724c844cddcb77298615a3c34d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4798569
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>