libdot: convert to ES6 module
Replace the generated index.js with an actual ES6 module that loads
all the rest of the hterm code on the fly.  This avoids the compile
step when testing locally.

This is structured as the minimal effort to convert libdot to an ES6
module.  We still have a global "lib" object that all the imported
files extend with their APIs.  This avoids a lot of disruption in
the tree and makes it easier to review.  This also avoids confusing
closure-compiler as it doesn't handle ES6 modules great where the
declared types use the full `lib.xxx` namespace.  Rectifying this
will take some effort, and perhaps writing custom externs, neither
of which is exactly great.

Change-Id: Id90c8206803b52e4c301ae84e33e4fed3807a850
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4801578
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>