hterm: convert to ES6 module
Replace the generated index.js with an actual ES6 module that loads
all the rest of the hterm code on the fly.  This avoids the compile
step when testing locally.

This is structured as the minimal effort to convert hterm to an ES6
module.  We still have a global "hterm" object that all the imported
files extend with their APIs.  This avoids a lot of disruption in
the tree and makes it easier to review.  This also avoids confusing
closure-compiler as it doesn't handle ES6 modules great where the
declared types use the full `hterm.xxx` namespace.  Rectifying this
will take some effort, and perhaps writing custom externs, neither
of which is exactly great.

Change-Id: Iaa071f6ad04677802e183f5cdeac5c4ef5ce8681
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4801579
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>