nassh: connect/popup: handle macOS command key better
Update the modifiers key logic to look for meta (command) key on macOS
and ctrl key everywhere else.  There isn't really a way to ask Chrome
what accelerators the current platform has bound, so this is about the
best we can do -- assume macOS is the only one to behave this way.

Bug: b/295572710
Change-Id: Ie7c64057594a761ff896be6d2bb8997c4ef59352
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4807268
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>