nassh: nasftp: support new File System Access API for saving large files
Newer versions of Chrome support the File System Access API which allows
us to get efficient writable handles to local files.  This allows us to
write directly to files when downloading instead of having to buffer the
entire file in memory before saving it.

The downside is there are more user interactions/dialogs that the user
has to go through when compared to the classic <a> interface (which has
no user prompts).  Because of that, continue using the old <a> method
when the file is "small" (arbitrarily picking a 10 MiB limit).

Bug: b/260529425
Change-Id: Id71ba315b5f21b9819e7b326980cc3e68c8a70db
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4837448
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>