wassh: have GAI return 2 results for AF_UNSPEC
Instead of assuming all requests are IPv6 or IPv4, and defaulting to
IPv4 if unspecified, return 2 results for all unspecified requests.
This isn't completely correct as it implies the host always has both
A & AAAA records, but it works in practice: since existence of records
doesn't mean we'll actually be able to connect to it, ssh will try to
connect to all the results until one works.  So we're just delaying
the failure until connect time where the missing record is handled.

While we're here, fix the numeric address parsing for AF_UNSPEC so
that we explicitly mark the result as IPv6 or IPv4.

Bug: b/308258000
Change-Id: Id923f4c4d79e5245618fb6424aa46dbccfef633a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/4988221
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>