node_modules: drop unused rollup-plugin-concat
The rollup-plugin-concat is no longer used, so let's punt it.
It was super helpful in migration from our concat to rollup, but
we've grown out of it.

More importantly, it hasn't been updated in a while, and it's
pulling in a lot of old deps that is holding others back.

Change-Id: I8b0468301cf82166e8e714b1800e427ce54f7635
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5003935
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>