wassh: fix IPv6 address parsing with socket get name
The sys_sock_get_name is expecting an array of 8-bits, not 16-bits.
IPv6 addresses are 16-bits separated by :, so we can't just split on
those.  Also, the Chrome socket APIs return compressed IPv6 addresses,
so we also have to expand those ourselves.  This should handle all of
the compressed forms like ::, ::1, 1::, 1::1, and uncompressed.

Change-Id: Ia96cfab1b5d213b6b8833c8dffe98c4634c2c146
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5006354
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>