libdot: prefs: switch PreferenceManager to a class
Now that we have ES6, cleanup our PreferenceManager classes.  This
helps closure-compiler type checking to find some places where we
had the wrong return types (in nassh's createProfile helpers).

This is a bit of a noisy diff, but it's largely reindenting existing
code and switching from standalone functions to class methods.

Change-Id: I9cb45a26c7fb6ab9a81f7dbe52c8b3104afa0b0a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5040193
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>