wassh: sockets: manually cleanup orphaned Chrome sockets
Since Chrome sockets are per-extension, not per-tab, sockets that are
not explicitly closed will leak.  This can happen if ssh isn't able to
cleanly close & exit.  For example, the user closes a window, or the
system suspends and the main connection hangs.  This is most noticeable
when using port forwarding as the local listening socket is left open.

Use the Chrome sockets's 'name' property to identify which tab a socket
belongs to.  When we start a new connection, walk all existing sockets
and see if their tabs are still open.  If not, close the socket.  Also,
if the socket is associated with the current tab, close it -- in case
we disconnect for some reason, and reconnect in the existing session.

Bug: b/294847208
Change-Id: I416c395dc2bc6977030cbafe228e7f3ac187d47e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5040199
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>