wassh: sockets: always set TCP keep alive interval with Chrome sockets
When enabling TCP keep alives with Chrome sockets, we enable the option,
but go with the default interval value (which is 0 seconds).  Chrome on
Linux (POSIX) will turn turn on SO_KEEPALIVE, but when delay is 0, skip
setting TCP_KEEPIDLE/TCP_KEEPINTVL, meaning the default OS values will
remain.  Chrome on Windows calls the kernel with delay of 0.  It seems
prior to Windows 11, this wasn't an issue, but now with Windows 11, this
is causing connections to get reset during initial setup.

Let's set the default to 75 seconds on all OS's and not rely on Chrome
managing it.

Thanks to Robert Raszuk & Maciej Żenczykowski for some deep debugging
to track this down.

Bug: b/303565674
Change-Id: Ibaa515ecd569cc61a70086424cc9e3766b2a9ae0
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5056706
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Emil Mikulic <easy@google.com>