libdot: prefs: fix restoring child ids list
When restoring a list of children from JSON, the process used to be:
* restore child 1 to memory & push to storage
* update list of children & push to storage
* restore child 2 to memory & push to storage
* update list of children & push to storage

The constant child list updating eats our storage write quota, so
we've been delaying the underlying storage update until the end:
* restore child 1 to memory & push to storage
* update list of children
* restore child 2 to memory & push to storage
* update list of children
* push list of children to storage

The set call itself optimizes storage interactions and doesn't write
if the in-memory value is the same as the new value it's being set to.
Since the list of children is up-to-date in-memory, trying to force
the final write via a redundant set call is skipped, and we end up not
writing the list to storage at all.

Workaround it for now by writing to storage directly.  This is a little
hacky, but the isImportingJson_ logic is a hack itself, and we'd need
to do a bit more work to clean it all up.

Bug: b/294126258
Change-Id: I2da36cda8092873f2c66f3aeec4c5793b7d45a6a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5075180
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>