libdot: isValidUrl: new helper to check URLs before use
When sometimes pass user-generated URIs to Web APIs, and if they are
invalid syntax, we don't want them crashing the app.  For example, a
user trying to Ctrl+Click arbitrary text they typed in hterm in order
to load that page.  Add a helper to check the syntax before we try to
use them.

Bug: b/329243940
Change-Id: Ic0ac58b103077f9d2001cf27eae7897cb99a335d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5374867
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>