nassh: support profile-id query string for selecting connection
Historically we've only supported connection selection via the hash in
the URI.  For example, .../nassh.html#profile-id:1234.  This causes us
problems as Chrome sometimes ignores the hash completely when creating
shortcuts on CrOS for pinning to the shelf.  It only supports unique
values in the query string.

We're already parsing the query string in the main page to pull out
other values, so we can support passing the profile-id via that too.
We can't drop support for old forms because (1) it'll break existing
bookmarks people have, (2) it's how we've documented & passed things
for over 10 years, and (3) it's how we've registered ssh:// URIs with
Chrome (as .../nassh.html#uri:xxx).

Bug: b/300669067
Change-Id: I1e060d1436a41bab276bd6eb48fd6fe8656c5e54
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5402414
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>