nassh: prefs: make it easy to save imported private keys
Private keys imported to Secure Shell do not have additional protections
on them -- if someone has access to your system, they can read the db
that the files are stored in.  Or they can open Secure Shell and paste a
little bit of JS to read the indexeddb directly.

With that in mind, make the identities clickable in the ssh prefs page.
This way people can easily extract any keys they imported in case they
deleted the keys locally.

Bug: b/40702444
Change-Id: Ia4800e187971d3490a533a78a1ab6cb9b7b7cb74
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5471402
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>