wassh: vfs: delay handle closing with duped file descriptors
When duping a fd, we don't want to close the underlying handle state
until all fds referencing that handle are closed.  We see this with
mosh that dupes a socket, closes the original fd, then uses the new
fd to send/recv network data.

The logic to see if there are still references isn't super efficient,
but the code isn't called that often, and we aren't expecting to have
a ton of open fds normally, so it's prob fine to be simple here until
we have a situation that calls for something more complicated.

Bug: b/40831504
Change-Id: I809dba2add2a57e96a7ce21254d4b2f4fc4b6cbc
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7134741
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>