ssh_client: wassh-libc-sup: implement very basic recvmsg
Since mosh uses recvmsg for network traffic, implement basic support
for it.  We assume only one buffer is used for data, and while we
provide control message structures & APIs, we don't actually return
any data in it since it's not really used by mosh for much.

For completeness, a sendmsg stub is provided too, but nothing seems
to use it currently.

Bug: b/40831504
Change-Id: I3b5ad11e6987daf09648b5dc3ce9a7d9d3907406
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5829885
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>