ssh_client: wasm: switch to wasi threads target
The default wasi target does not support any atomic or pthread APIs,
even ones that would assume a single thread.  Those APIs are only
available in the new experimental wasi threads target.  While the
OpenSSH code doesn't use them, mosh (and its protobuf dep) rely on
them, so switch over to make it work.

This results in slightly larger code at runtime, but perf seems to
be about the same.

We can't compile just the mosh code with pthreads because all libraries
have to use the same target (otherwise you get link errors or weird
runtime misbehavior), and OpenSSH & mosh share a bunch of libraries.
With more effort, we could compile all the libraries twice (once for
pthreads & once for non-pthreads), but that seems like a lot of extra
build time overhead when the runtime impact is negligible.

Bug: b/40831504
Change-Id: Ia74f030be70dbf545311120882f23cb066e244f5
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/5838169
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>