wasi-js-bindings: switch Program.run to async
The new JSPI for WASM allows programs to call async functions, but in
order to do so, the entry point needs to be async as well.  Adjust the
run API to always be async.  This change by itself doesn't enable the
new JSPI, but merely adjusts the public API so we know it's safe.

Change-Id: I068e5264806073ebf91478deb1b69fc1b3f23c1d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7168661
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>