ssh_client: wassh-libc-sup: stub out closefrom
Rather than patch out one call to closefrom in the ssh program, stub
out the call for everyone.  OpenSSH uses it in many places, and they
would all break the WASI runtime as a result.  We largely don't have
to worry about the fd leakage that this function protects against as
every WASM process starts up with a pristine environment.

Change-Id: Ia588bb331936503ebbf70d76aa66e7bd3a865f45
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7036558
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>