wassh: increase syscall buffer size to 256KiB
Starting with OpenSSH 8.9, ssh can perform reads on up to 256KiB buffers
where previously it would top out at like 16KiB.  The shared buffer used
to pass results between the processes was set to 64KiB which would run
out of space on the larger reads.

We should plumb the max buffer size down to the handlers so they don't
try to return more data than can be handled, but that'll take more work,
and we need to unblock current OpenSSH builds.

Bug: b/452419036
Bug: b/295308963
Change-Id: If422749a5e4fa01691fdade39c1cfe7ebe83ad96
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7042057
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>