ssh_client: rework library installs into ABI-specific dirs
We've been installing WASM builds into a single path.  While this
made sense initially as WASI only had 1 ABI, now that it's expanding
into multiple ones that we want to support simultaneously, we need
to switch our installs to be ABI-specific.

The way it works today:
  output/wasi-sdk/share/wasi-sysroot/
    lib/
      libz.a
    include/
      zlib.h

Since we only support WASI preview1 (single threaded), there are no
conflicts.  But if we want to also build for WASI preview1 (threaded),
the libz.a files clobber each other.  This is needed by mosh as it is
written in C++ and uses exceptions which are only supported by the new
threaded builds.  We tried to use the threaded variant for OpenSSH too
but it generated larger code & crashed at runtime.

Upstream WASI already reworked their sysroot layout:
  output/wasi-sdk/share/wasi-sysroot/
    lib/
      wasm32-wasip1/
        libc.a
      wasm32-wasip1-threads/
        libc.a
    include/
      wasm32-wasip1/
        stdio.h
      wasm32-wasip1-threads/
        stdio.h

Which means switching paths is mostly straightforward -- we put the
ABI name under the target dirs too.
  output/wasi-sdk/share/wasi-sysroot/
    lib/
      wasm32-wasip1/
        libz.a
    include/
      wasm32-wasip1/
        zlib.h

The only pain point here is OpenSSL (and LDNS).  OpenSSL doesn't
allow custom include paths (it hardocdes <prefix>/include/openssl/),
and LDNS has very crufty build logic to snoop the filesystem instead
of querying the toolchain via compile+link tests.  We should see if
we can improve these upstreams eventually.

This will also make it easier to migrate to newer WASI APIs if we
decide to as we can build & ship both simultaneously.  We don't have
immediate plans to support preview2, but considering it has added a
lot more network APIs, we should investigate at some point.

Bug: b/40831504
Change-Id: I56c77090acac65b4bbb5bd09921ad07beb47207f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7047326
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>