nassh: release: fix release notes display
After reworking the note display with rollup, the format changed
slightly, but it ended up working correctly because there was only
one entry.  Now that we need to add more, the output is incorrect
with how newlines are inserted.  Rework it to support the new array
format generated from the yaml file.

Change-Id: Idd1c0c7526f06308755f2305d43262fb45e28e93
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7092234
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>