nassh: uri: allow -oEnableEscapeCommandline=yes
The OpenSSH 9.2 release changed the default behavior with ~C to disable
it.  Allow URIs to override and re-enable since, from the client POV,
this isn't a security issue.

This is a bit of a hack in special casing this specific option in this
specific form, but we want to unblock the folks who are relying on this
and I don't really want to rollback OpenSSH versions over it.

Bug: 457316822
Change-Id: I3b5e7ad20ca5bac2373ddb96826f5fe191063b5f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7116019
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>