ssh_client: increase stacksize significantly
OpenSSH 10.0 fails during startup with memory access out of bounds
errors.  Turns out, we're blowing the stack because the default size
is a mere 64KiB.  I'm surprised we hadn't exceeded this sooner.  But
if we haven't received any complaints in the years since the initial
WASM launch, let's double it and call it a day.

Url: https://github.com/llvm/llvm-project/issues/151015
Change-Id: I41c4b040bbecc35e8acd85e924d0629945c31fcc
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7127784
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>