nassh: connect: rework app UI selection
Rather than control which app to run by an ever increasing set of
buttons at the bottom, turn it into a drop down list.  This has the
side-effect of making saved connections per-app, but this is a good
thing as it'll allow people to leverage these different modes via
the extension popup too whereas today they can only start ssh.  It
also avoids us having to filter ssh options between different modes
and try to guess which are appropriate.  For example, if someone
has an ssh connection with port forwarding, do they want to enable
those when connecting via sftp?

Bug: b/40831504
Change-Id: I19d9c284f2eb43ae50b77d089f145e385e886830
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7494013
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>