nassh: connect: move app setup to command instance
Currently the connect dialog passes along the app type to the command
instance via session storage.  This is a bit wonky, and it means you
can't pin the app in the connection and then load that connection via
other means (e.g. bookmarks) -- it'll always default to ssh.  If we
move this logic out of the connect dialog, it makes things simpler and
easier to reason about (he connect dialog is purely about choosing a
saved connection profile), and avoids session indirection weirdness.

Change-Id: I83e33c5ce11582634257e7c832ef62cc56049295
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7494014
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>