nassh: browserAction: drop all usage
Since we define a popup (via the manifest), the onClicked event never
fires, which makes all of this logic a bit pointless.  If we wanted to
change badge text or something, there might be value, but since there
are no plans along those lines, let's just drop all the code.

Change-Id: Ie909e615209a19fe67e1a7a7d2448eac41699635
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7157403
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>