nassh: subproc: rename "plugin" to "subproc"
The "plugin" nomenclature is tied to the HTML runtime and how NaCl was
an <embed> element added to the DOM with a src attribute pointing to the
plugin manifest.  Now that NaCl is gone, and we're centered around WASM
which only uses JS APIs, keeping "plugin" is mismatched with the WASI
naming conventions.  The WJB & wassh layers already use "program" to
represent a compiled program, and on top of that is a "process" that
represents a runtime.  The JS world often uses "app" for a single web
page or major UI element, so that term is out.  Let's use "subproc"
as a low-level term akin to the Python library.  This will hold very
basic logic on top of specific programs that we want to run which
should make it easier to integrate more programs (e.g. mosh) into the
project without blowing out existing modules even more.

Bug: b/40831504
Change-Id: I0f624a12fb29a5211ccaf577e67d320cba262c63
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7494010
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>