nassh: streams: drop asyncWrite callback
With the NaCl removal, nothing uses the callback property anymore, so
drop it to simplify.  It was always a bit of a janky interface because
it assumed the callback was always viable, and you could only ever have
one set -- whichever function last wrote to it.  That meant code had to
write and wait for the callback before it could possibly write more.

Change-Id: I5ab1c629dd73b417548776df06115aa088930873
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7195221
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>