nassh: streams: change asyncWrite into async
The only caller already assumes it returns a promise, so changing it
to async is basically a no-op.  It doesn't have a callback, so there's
no real work to do in the implementations anymore.  Some of the funcs
could be adjusted to use setTimeout less, but best to leave that as a
followup so it can be better tested.

Change-Id: I0a07dc7638c2dcf7aa9e528f7db3bc2cf8b9c3ed
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7195222
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>