wasi-js-bindings: add helper for encoding into SharedArrayBuffers
Since WASM modules might use SharedArrayBuffers for their memory, we
can't assume encodeInto will work as it only supports ArrayBuffers
under Chrome currently.  Extract the logic from SyscallLock into a
dedicated helper to dynamically handle this situation, and use that
in all the places we currently use encodeInto.  If the runtime allows
it, we'll do the fast path (encoding directly into the memory), but
if it fails, we'll automatically fallback to encoding into a buffer
and copying that back.

Change-Id: I9979a5beaf85169906cb33c1e3687a2c0b42b6ab
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/7205701
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>